backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.26.3
python-multipart==0.0.6
pypdf2==3.0.1
markdown==3.5.2
pydantic==2.5.3
aiofiles==23.2.1
transformers==4.37.2
torch==2.1.2
